Fixed silly reconcile_bucket method.  Much simpler now.
Instead of trying to work with just adjacent buckets, work with all buckets lesser than or greater than the current bucket when sorting.  Then merge the "during" bits and everything should be easy.
Keeping track of the unsorted lists is much easier this way too: if a photo is sorted in one bucket, it's automatically sorted in all buckets greater than the given bucket.

Also added some more complicated tests.  Hopefully this thing is good to go now.